ShoppingCartMS/requirements.txt
Flask
Flask-Migrate
connexion[swagger-ui]
sqlalchemy
pyjwt
marshmallow-sqlalchemy